// src/app/precisa/ventas/Contrato/model/icontrato.ts
export interface IContrato {
    Nro:string;

    DOCUMENTO_CLIENTE: string;
    CLIENTE: string;
    DOCUMENTO_CONYUGE: string;
    CONYUGE: string;
    COMPAÑIA: string;
    SUCURSAL: string;
    PROGRAMA: string;
    MANZANA: string;
    LOTE: string;

    AREA_M2: string;
    PRECIO_M2: string;

    //COSTO_TOTAL: string;
    TIPO_CAMBIO: string;
    MONTO_SEPARACION: string;
    TIPO_MONEDA: string;

    CUOTA_INICIAL: string;
    NRO_CUOTA: string;
    CUOTA_MENSUAL: string;
    CUOTA_FINAL: string;

    FECHA_INICIAL: string;
    FECHA_PRIMERA_LETRA: string;
    // DIAS_GRACIA: string;
    SALDO: string;
    // RETENCION: string;
    TASA_MORA: string;
    OBSERVACION: string;

    DOCUMENTO_COMISIONISTA: string;
    COMISIONISTA: string;
    MONTO_TOTAL: string;



}