// src/app/precisa/proyecto/Lotes/model/FiltroLote.ts
export class FiltroLote{

    constructor(){

    }
    IdLote:         number;
    IdManzana:      number;
    IdProyecto:     number;
    IdSede:         number;
    CompaniaCodigo: string;
    Codigo:         string;
    Nombre:         string;
    Estado:         number;
    Departamento:   string;
    Provincia:      string;
    Distrito:       string;    
}